// libs/nx-aws-s3-code-deploy/src/lib/deploy/schema.d.ts
export interface Schema {
  configuration?: string;
  buildTarget?: string;
  noBuild?: string;
  baseHref?: string;
  accessKeyId?: string;
  secretAccessKey?: string;
  region?: string;
  bucket?: string;
  zip?: string;
  deploymentGroupName?: string;
  applicationName?: string;
  subFolder?: string;
}